Feature: Reorganise pages and add footer for each page.
